refactor: move to json serializable request
